Policy restructure (#134)
* moved actor critic policies to base_policies directory

* Rendered TeX expressions in 07b813e692ed70d42dc9b1d564be65b7e95a8951

* moved layer to tf_util

* moved policy tests by submodules

* bug fix

* added docstring
