+++ hbaselines/base_policies/__init__.py
@@ -1,0 +1,6 @@
+"""Init file for the base policies submodule."""
+from hbaselines.base_policies.actor_critic import ActorCriticPolicy
+
+__all__ = [
+    "ActorCriticPolicy"
+]
