+++ hbaselines/fcnet/sac.py
@@ -2,8 +2,9 @@
 import tensorflow as tf
 import numpy as np
 
-from hbaselines.fcnet.base import ActorCriticPolicy
+from hbaselines.base_policies import ActorCriticPolicy
 from hbaselines.fcnet.replay_buffer import ReplayBuffer
+from hbaselines.utils.tf_util import layer
 from hbaselines.utils.tf_util import get_trainable_vars
 from hbaselines.utils.tf_util import reduce_std
 from hbaselines.utils.tf_util import gaussian_likelihood
@@ -362,14 +363,14 @@ class FeedForwardPolicy(ActorCriticPolicy):
 
             # create the hidden layers
             for i, layer_size in enumerate(self.layers):
-                pi_h = self._layer(
+                pi_h = layer(
                     pi_h,  layer_size, 'fc{}'.format(i),
                     act_fun=self.act_fun,
                     layer_norm=self.layer_norm
                 )
 
             # create the output mean
-            policy_mean = self._layer(
+            policy_mean = layer(
                 pi_h, self.ac_space.shape[0], 'mean',
                 act_fun=None,
                 kernel_initializer=tf.random_uniform_initializer(
@@ -377,7 +378,7 @@ class FeedForwardPolicy(ActorCriticPolicy):
             )
 
             # create the output log_std
-            log_std = self._layer(
+            log_std = layer(
                 pi_h, self.ac_space.shape[0], 'log_std',
                 act_fun=None,
             )
@@ -453,14 +454,14 @@ class FeedForwardPolicy(ActorCriticPolicy):
 
                     # create the hidden layers
                     for i, layer_size in enumerate(self.layers):
-                        vf_h = self._layer(
+                        vf_h = layer(
                             vf_h, layer_size, 'fc{}'.format(i),
                             act_fun=self.act_fun,
                             layer_norm=self.layer_norm
                         )
 
                     # create the output layer
-                    value_fn = self._layer(
+                    value_fn = layer(
                         vf_h, 1, 'vf_output',
                         kernel_initializer=tf.random_uniform_initializer(
                             minval=-3e-3, maxval=3e-3)
@@ -476,14 +477,14 @@ class FeedForwardPolicy(ActorCriticPolicy):
 
                     # create the hidden layers
                     for i, layer_size in enumerate(self.layers):
-                        qf1_h = self._layer(
+                        qf1_h = layer(
                             qf1_h, layer_size, 'fc{}'.format(i),
                             act_fun=self.act_fun,
                             layer_norm=self.layer_norm
                         )
 
                     # create the output layer
-                    qf1 = self._layer(
+                    qf1 = layer(
                         qf1_h, 1, 'qf_output',
                         kernel_initializer=tf.random_uniform_initializer(
                             minval=-3e-3, maxval=3e-3)
@@ -495,14 +496,14 @@ class FeedForwardPolicy(ActorCriticPolicy):
 
                     # create the hidden layers
                     for i, layer_size in enumerate(self.layers):
-                        qf2_h = self._layer(
+                        qf2_h = layer(
                             qf2_h, layer_size, 'fc{}'.format(i),
                             act_fun=self.act_fun,
                             layer_norm=self.layer_norm
                         )
 
                     # create the output layer
-                    qf2 = self._layer(
+                    qf2 = layer(
                         qf2_h, 1, 'qf_output',
                         kernel_initializer=tf.random_uniform_initializer(
                             minval=-3e-3, maxval=3e-3)
