+++ hbaselines/multi_fcnet/sac.py
@@ -6,6 +6,7 @@ from hbaselines.multi_fcnet.base import MultiFeedForwardPolicy as BasePolicy
 from hbaselines.fcnet.sac import FeedForwardPolicy
 from hbaselines.multi_fcnet.replay_buffer import MultiReplayBuffer
 from hbaselines.multi_fcnet.replay_buffer import SharedReplayBuffer
+from hbaselines.utils.tf_util import layer
 from hbaselines.utils.tf_util import get_trainable_vars
 from hbaselines.utils.tf_util import reduce_std
 from hbaselines.utils.tf_util import gaussian_likelihood
@@ -801,14 +802,14 @@ class MultiFeedForwardPolicy(BasePolicy):
 
             # create the hidden layers
             for i, layer_size in enumerate(self.layers):
-                pi_h = self._layer(
+                pi_h = layer(
                     pi_h,  layer_size, 'fc{}'.format(i),
                     act_fun=self.act_fun,
                     layer_norm=self.layer_norm
                 )
 
             # create the output mean
-            policy_mean = self._layer(
+            policy_mean = layer(
                 pi_h, ac_space.shape[0], 'mean',
                 act_fun=None,
                 kernel_initializer=tf.random_uniform_initializer(
@@ -816,7 +817,7 @@ class MultiFeedForwardPolicy(BasePolicy):
             )
 
             # create the output log_std
-            log_std = self._layer(
+            log_std = layer(
                 pi_h, ac_space.shape[0], 'log_std',
                 act_fun=None,
             )
@@ -883,14 +884,14 @@ class MultiFeedForwardPolicy(BasePolicy):
 
                     # create the hidden layers
                     for i, layer_size in enumerate(self.layers):
-                        vf_h = self._layer(
+                        vf_h = layer(
                             vf_h, layer_size, 'fc{}'.format(i),
                             act_fun=self.act_fun,
                             layer_norm=self.layer_norm
                         )
 
                     # create the output layer
-                    value_fn = self._layer(
+                    value_fn = layer(
                         vf_h, 1, 'vf_output',
                         kernel_initializer=tf.random_uniform_initializer(
                             minval=-3e-3, maxval=3e-3)
@@ -906,14 +907,14 @@ class MultiFeedForwardPolicy(BasePolicy):
 
                     # create the hidden layers
                     for i, layer_size in enumerate(self.layers):
-                        qf1_h = self._layer(
+                        qf1_h = layer(
                             qf1_h, layer_size, 'fc{}'.format(i),
                             act_fun=self.act_fun,
                             layer_norm=self.layer_norm
                         )
 
                     # create the output layer
-                    qf1 = self._layer(
+                    qf1 = layer(
                         qf1_h, 1, 'qf_output',
                         kernel_initializer=tf.random_uniform_initializer(
                             minval=-3e-3, maxval=3e-3)
@@ -925,14 +926,14 @@ class MultiFeedForwardPolicy(BasePolicy):
 
                     # create the hidden layers
                     for i, layer_size in enumerate(self.layers):
-                        qf2_h = self._layer(
+                        qf2_h = layer(
                             qf2_h, layer_size, 'fc{}'.format(i),
                             act_fun=self.act_fun,
                             layer_norm=self.layer_norm
                         )
 
                     # create the output layer
-                    qf2 = self._layer(
+                    qf2 = layer(
                         qf2_h, 1, 'qf_output',
                         kernel_initializer=tf.random_uniform_initializer(
                             minval=-3e-3, maxval=3e-3)
