+++ hbaselines/utils/tf_util.py
@@ -1,5 +1,6 @@
 """TensorFlow utility methods."""
 import tensorflow as tf
+import tensorflow.contrib.slim as slim
 import numpy as np
 from functools import reduce
 
@@ -222,3 +223,44 @@ def print_params_shape(scope, param_type):
     nb_params = sum([reduce(lambda x, y: x * y, shape) for shape in shapes])
     print('  {} shapes: {}'.format(param_type, shapes))
     print('  {} params: {}'.format(param_type, nb_params))
+
+
+def layer(val,
+          num_outputs,
+          name,
+          act_fun=None,
+          kernel_initializer=slim.variance_scaling_initializer(
+              factor=1.0 / 3.0, mode='FAN_IN', uniform=True),
+          layer_norm=False):
+    """Create a fully-connected layer.
+
+    Parameters
+    ----------
+    val : tf.Variable
+        the input to the layer
+    num_outputs : int
+        number of outputs from the layer
+    name : str
+        the scope of the layer
+    act_fun : tf.nn.* or None
+        the activation function
+    kernel_initializer : Any
+        the initializing operation to the weights of the layer
+    layer_norm : bool
+        whether to enable layer normalization
+
+    Returns
+    -------
+    tf.Variable
+        the output from the layer
+    """
+    val = tf.layers.dense(
+        val, num_outputs, name=name, kernel_initializer=kernel_initializer)
+
+    if layer_norm:
+        val = tf.contrib.layers.layer_norm(val, center=True, scale=True)
+
+    if act_fun is not None:
+        val = act_fun(val)
+
+    return val
